fix: resolve CI failures - remove unused exports and refactor to 2 modules
- Remove unused exported types to fix knip errors
- Fix async test for lazy loading modal
- Refactor from 3 modules to 2 (search.ts and search-ui.ts)
- Combine SearchAPI class into search-ui.ts since they're always used together
